refactor: moved model classes to model directory
